Move structs from vga.h & render.h to video.h
