+++ include/render.h
@@ -46,81 +46,6 @@ struct RenderPal_t {
 	uint32_t last         = 0;
 };
 
-enum class PixelFormat : uint8_t {
-	// Up to 256 colours, paletted;
-	// stored as packed uint8 data
-	Indexed8 = 8,
-
-	// 32K high colour, 5 bits per red/blue/green component;
-	// stored as packed uint16 data with highest bit unused
-	BGR555 = 15,
-	//
-	// 65K high colour, 5 bits for red/blue, 6 bit for green;
-	// stored as packed uint16 data
-	BGR565 = 16,
-	//
-	// 16.7M (24-bit) true colour, 8 bits per red/blue/green component;
-	// stored as packed 24-bit data
-	BGR888 = 24,
-	//
-	// 16.7M (32-bit) true colour; 8 bits per red/blue/green component;
-	// stored as packed uint32 data with lowest 8 bits unused
-	BGRX8888 = 32
-};
-
-const char* to_string(const PixelFormat pf);
-
-uint8_t get_bits_per_pixel(const PixelFormat pf);
-
-// Details about the rendered image.
-// E.g. for the 320x200 256-colour 13h VGA mode with double-scanning
-// and pixel-doubling enabled:
-//   - width = 320 (will be pixel-doubled post-render via double_width)
-//   - height = 400 (2*200 lines because we're rendering scan-doubled)
-//   - pixel_aspect_ratio = 5/6 (1:1.2) (because the PAR is meant for the
-//     final image, post the optional width & height doubling)
-//   - double_width = true (pixel-doubling)
-//   - double_height = false (we're rendering scan-doubled)
-//
-struct RenderParams {
-	// Width of the rendered image (prior to optional width-doubling)
-	uint16_t width = 0;
-
-	// Height of the rendered image (prior to optional height-doubling)
-	uint16_t height = 0;
-
-	// If true, the rendered image is doubled horizontally after via
-	// a scaler (e.g. to achieve pixel-doubling)
-	bool double_width = false;
-
-	// If true, the rendered image is doubled vertically via a
-	// scaler (e.g. to achieve fake double-scanning)
-	bool double_height = false;
-
-	// If true, we're dealing with a double-scanned VGA mode that was forced
-	// to be rendered as single-scanned
-	bool force_single_scan = false;
-
-	// Pixel aspect ratio to be applied to the final image (post
-	// width & height doubling) so it appears as intended on screen.
-	// (video_mode.pixel_aspect_ratio holds the "nominal" pixel
-	// aspect ratio of the source video mode)
-	Fraction pixel_aspect_ratio = {};
-
-	// Pixel format of the image data
-	PixelFormat pixel_format = {};
-
-	// Details about the source video mode.
-	// This is usually different than the rendered image's details.
-	// E.g. for the 320x200 256-colour 13h VGA mode we always have the
-	// following, regardless of whether double-scanning and pixel-doubling
-	// is enabled at the rendering level:
-	//   - width = 320
-	//   - height = 200
-	//   - pixel_aspect_ratio = 5/6 (1:1.2)
-	VideoMode video_mode = {};
-};
-
 struct Render_t {
 	RenderParams src = {};
 	uint32_t src_start   = 0;
